Move and rename locationApi.ts
Same as the other two
